fix migration from 0.3.5.1 (#13)

+++ build/scripts/migration/script.d/01-migrate-gateway.sh
@@ -20,6 +20,17 @@ __error() {
     exit 1
 }
 
+__normalize_version() {
+    local version
+    if [ "${1::1}" = "v" ]; then
+        version="${1:1}"
+    else
+        version="${1}"
+    fi
+
+    echo "$version"
+}
+
 __is_version_gt() {
     test "$(echo "$@" | tr " " "\n" | sort -V | head -n 1)" != "$1"
 }
@@ -28,8 +39,8 @@ __is_migration_needed() {
     local version1
     local version2
 
-    version1="${1}"
-    version2="${2}"
+    version1=$(__normalize_version "${1}")
+    version2=$(__normalize_version "${2}")
 
     if [ "${version1}" = "${version2}" ]; then
         return 1
@@ -76,7 +87,12 @@ if [ "${NEED_MIGRATION}" = "false" ]; then
     exit 0
 fi
 
-MIGRATION_SERVICE_DIR=${BUILD_PATH}/scripts/migration/service.d/${APP_NAME_SHORT}
+MIGRATION_SERVICE_DIR=${1}
+
+if [ -z "${MIGRATION_SERVICE_DIR}" ]; then
+    MIGRATION_SERVICE_DIR=${BUILD_PATH}/scripts/migration/service.d/${APP_NAME_SHORT}
+fi
+
 MIGRATION_LIST_FILE=${MIGRATION_SERVICE_DIR}/migration.list
 MIGRATION_PATH=()
 
@@ -97,7 +113,7 @@ while read -r VERSION_PAIR; do
     # obtain "v0.3.6-alpha2" from "v0.3.5 v0.3.6-alpha2"
     VER2=$(echo "${VERSION_PAIR}" | cut -d' ' -f2)
 
-    if [ "v${CURRENT_VERSION}" = "${VER1// /}" ] || [ "${CURRENT_VERSION}" = "LEGACY_WITHOUT_VERSION" ]; then
+    if [ "${CURRENT_VERSION}" = "${VER1// /}" ] || [ "${CURRENT_VERSION}" = "LEGACY_WITHOUT_VERSION" ]; then
         CURRENT_VERSION_FOUND="true"
     fi
 
@@ -132,6 +148,13 @@ pushd "${MIGRATION_SERVICE_DIR}"
 
 {
     for VER2 in "${MIGRATION_PATH[@]}"; do
+        MIGRATION_TOOL_FILE=linux-"${ARCH}"-"${APP_NAME}"-migration-tool-"${VER2}".tar.gz
+
+        if [ -f "${MIGRATION_TOOL_FILE}" ]; then
+            __info "Migration tool ${MIGRATION_TOOL_FILE} exists. Skip downloading."
+            continue
+        fi
+
         MIGRATION_TOOL_URL=https://github.com/IceWhaleTech/"${APP_NAME_FORMAL}"/releases/download/"${VER2}"/linux-"${ARCH}"-"${APP_NAME}"-migration-tool-"${VER2}".tar.gz
         __info "Dowloading ${MIGRATION_TOOL_URL}..."
         curl -sL -O "${MIGRATION_TOOL_URL}"
@@ -146,9 +169,6 @@ pushd "${MIGRATION_SERVICE_DIR}"
         MIGRATION_TOOL_FILE=linux-"${ARCH}"-"${APP_NAME}"-migration-tool-"${VER2}".tar.gz
         __info "Extracting ${MIGRATION_TOOL_FILE}..."
         tar zxvf "${MIGRATION_TOOL_FILE}" || __error "Failed to extract ${MIGRATION_TOOL_FILE}"
-
-        MIGRATION_SYSROOT_DIR=$(realpath -e "${MIGRATION_SERVICE_DIR}"/build/sysroot || __error "Failed to find sysroot directory for migration")
-        cp -rv "${MIGRATION_SYSROOT_DIR}"/* / || __error "Failed to copy sysroot directory for migration"
 
         MIGRATION_TOOL_PATH=build/sysroot/usr/bin/${APP_NAME}-migration-tool
         __info "Running ${MIGRATION_TOOL_PATH}..."
